+++ cmem_plugin_shapes/__init__.py
@@ -1,1 +1,1 @@
-"""shapes - main package"""
+"""cmem-plugin-shapes"""