+++ update_prefixes.py
@@ -7,5 +7,5 @@ from urllib.request import urlopen
 with urlopen("http://prefix.cc/popular/all.file.json") as remote_file:
     json_data = loads(remote_file.read())
 
-with (Path("cmem_plugin_shapes") / "prefix.cc.json").open("w") as local_file:
+with (Path("cmem_plugin_shapes") / "prefix_cc.json").open("w") as local_file:
     dump(json_data, local_file, sort_keys=True, indent=2)