+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [unreleased] 
+
+### Changed
+
+- removed dependency str2bool
+- use GraphParameterType for Shapes graph 
+
 ## [1.0.0beta1] 2024-07-24
 
 ### Added